backend-original/requirements.txt
flask
flask-cors
flask-pymongo
pymongo[srv]
python-dotenv
pyjwt
bcrypt